Refactor mock logic into separate files
