+++ pallets/vault-registry/src/mock.rs
@@ -18,7 +18,7 @@ pub use currency::testing_constants::{
 };
 use oracle::{
 	dia::DiaOracleAdapter,
-	oracle_mock::{
+	testing_utils::{
 		MockConvertMoment, MockConvertPrice, MockDataCollector, MockDiaOracle,
 		MockOracleKeyConvertor,
 	},
